Renamed thread.py to manager.py
