+++ src/SerialPlotter/manager.py
